+++ CHANGELOG.md
@@ -49,6 +49,14 @@ x.x.x Release notes (yyyy-MM-dd)
   correctly nil out nullable properties when updating an existing
   object when the `value` argument specifies nil or `NSNull` for
   the property value.
+* Computed properties on Realm object types are detected and no
+  longer added to the automatically generated schema.
+* Remove deprecated `{RLM}SyncPermission` and `{RLM}SyncPermissionChange`
+  classes. Use `{RLM}SyncPermissionValue` instead.
+* Remove deprecated `-[RLMSyncUser permissionRealmWithError:]` and
+  `SyncUser.permissionRealm()` APIs. Use the new permissions system.
+* Remove deprecated error `RLMSyncErrorBadResponse`. Use
+  `RLMSyncAuthErrorBadResponse` instead.
 
 ### Enhancements